+++ src/vec3.h
@@ -7,12 +7,12 @@
 namespace raytracer {
 class vec3 {
    public:
-    double x, y, z;
+    float x, y, z;
 
    public:
     vec3() : x(0), y(0), z(0) {}
-    vec3(double v) : x(v), y(v), z(v) {}
-    vec3(double x0, double y0, double z0 = 0) : x(x0), y(y0), z(z0) {}
+    vec3(float v) : x(v), y(v), z(v) {}
+    vec3(float x0, float y0, float z0 = 0) : x(x0), y(y0), z(z0) {}
     vec3(const vec3& other) = default;
     vec3& operator=(const vec3& other) {
         if (this == &other)
@@ -23,18 +23,18 @@ class vec3 {
         return *this;
     }
 
-    vec3 operator*(double a) const { return vec3(x * a, y * a, z * a); }
+    vec3 operator*(float a) const { return vec3(x * a, y * a, z * a); }
     vec3 operator*(const vec3 r) const { return vec3(x * r.x, y * r.y, z * r.z); }
-    vec3 operator/(const double r) const { return fabs(r) > epsilon ? vec3(x / r, y / r, z / r) : vec3(0, 0, 0); }
+    vec3 operator/(const float r) const { return fabs(r) > epsilon ? vec3(x / r, y / r, z / r) : vec3(0, 0, 0); }
     vec3 operator+(const vec3& v) const { return vec3(x + v.x, y + v.y, z + v.z); }
     vec3 operator-(const vec3& v) const { return vec3(x - v.x, y - v.y, z - v.z); }
     vec3 operator-() const { return vec3(-x, -y, -z); }
     void operator+=(const vec3& v) { x += v.x, y += v.y, z += v.z; }
-    void operator*=(double a) { x *= a, y *= a, z *= a; }
+    void operator*=(float a) { x *= a, y *= a, z *= a; }
     void operator*=(const vec3& v) { x *= v.x, y *= v.y, z *= v.z; }
     bool operator==(const vec3& v) const { return (fabs(x - v.x) < epsilon && fabs(y - v.y) < epsilon && fabs(z - v.z) < epsilon); }
-    double length() const { return sqrt(x * x + y * y + z * z); }
-    double average() { return (x + y + z) / 3; }
+    float length() const { return sqrt(x * x + y * y + z * z); }
+    float average() { return (x + y + z) / 3; }
     vec3 normalize() const { return (*this) / length(); }
 };
 
@@ -42,7 +42,7 @@ using point3 = vec3;  // 3D point
 
 // Vec3 Utility Functions
 
-inline double dot(const vec3& v1, const vec3& v2) {
+inline float dot(const vec3& v1, const vec3& v2) {
     return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z);
 }
